requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.4
gunicorn==21.2.0
h11==0.14.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
whitenoise==6.6.0
